order-form

+++ src/components/Order/Order.js
@@ -3,6 +3,9 @@ import {useCartContext} from '../../contexts/CartContext';
 import {useHistory} from 'react-router-dom';
 import { getFirestore } from '../../services/getProducts';
 import firebase from 'firebase';
+import './Order.css';
+import headerPhoto from '../../assets/img/bourne-style.jpg';
+import Swal from 'sweetalert2';
 
 const Order = () => {
 
@@ -32,35 +35,45 @@ const Order = () => {
             const title = cartItem.item.title;
             const price = cartItem.item.price * cartItem.cantidad;
             const image = cartItem.item.images[0];
+            const quantity = cartItem.cantidad;
 
-            return { id, title, price, image };
+            return { id, title, price, image, quantity };
         })
         const db = getFirestore();
         const ordersCollection = db.collection('orders') // creo nueva colección
         ordersCollection.add(order)
             .then((idDocumento) => {
-                alert(`Su orden ${idDocumento.id} está siendo procesada.`);
-                handleLink(`order/${idDocumento.id}`);
+                Swal.fire({
+                    title: '¿Estás seguro?',
+                    text: "¿Deseas confirmar la compra?",
+                    icon: 'question',
+                    showCancelButton: true,
+                    confirmButtonColor: '#3085d6',
+                    cancelButtonColor: '#d33',
+                    confirmButtonText: 'Confirmar'
+                  }).then((result) => {
+                    if (result.isConfirmed) {
+                      Swal.fire(
+                        '¡Muchas gracias por confiar en nosotros!',
+                        `Tu orden de compra es ${idDocumento.id}`,
+                        'success',
+                      )
+                      setBuyer({
+                        name:'',
+                        phone:'',
+                        email: ''
+                    });
+                    deleteCart();
+                    handleLink(`order/${idDocumento.id}`);
+                    }
+                  });
             })
             .catch((err) => {
                 console.log(err);
-            })
-            .finally(()=>{
-                alert('Su compra ha finalizado de manera exitosa');
-                setBuyer({
-                    name:'',
-                    phone:'',
-                    email: ''
-                });
             });
-
-            deleteCart();
     }
 
-    const pxq = (a,b) => {
-        return a*b
-      }
-      let total = 0;
+    let total = 0;
     
     const totalPxQ = (a, b) => {
         let sum = a*b;
@@ -79,21 +92,42 @@ const Order = () => {
         })
     }
     return (
-            <div className='modal-container'>
-                    <h2>Datos del comprador</h2>
-                    <form onSubmit={generateOrder}>
-                        <div>
-                            <input type='text' onChange={handleChange}  name='name' placeholder='name' value={buyer.name}/>
-                        </div>
-                        <div>
-                            <input type='text' onChange={handleChange}  name='phone'placeholder='tel' value={buyer.phone}/>
-                        </div>
-                        <div>
-                            <input type='email' onChange={handleChange}  name='email'placeholder='email' value={buyer.email}/>
-                        </div>
-                <button>Comprar</button>
-            </form>
-    </div>
+        <>
+        {!cart.length ? (
+            <div className='order-container margin-b'>
+                <div className='img-form'>
+                    <img src={headerPhoto} alt='form'/>
+                    <h2>No hay elementos en el carrito</h2>
+                    <button onClick={() => handleLink('/products')}>Ir al catálogo</button>
+                </div>
+            </div>
+            ) 
+            : 
+            (
+            <div className='order-container'>
+                <div className='img-form'>
+                    <img src={headerPhoto} alt='form'/>
+                </div>
+                <h2>Datos del comprador</h2>
+                <form onSubmit={generateOrder}>
+                    <div>
+                        <h3>Nombre</h3>
+                        <input type='text' onChange={handleChange}  name='name' placeholder='Nombre' value={buyer.name}/>
+                    </div>
+                    <div>
+                        <h3>Teléfono</h3>
+                        <input type='text' onChange={handleChange}  name='phone'placeholder='Teléfono' value={buyer.phone}/>
+                    </div>
+                    <div>
+                        <h3>Email</h3>
+                        <input type='email' onChange={handleChange}  name='email' placeholder='Email' value={buyer.email}/>
+                    </div>
+                    <button disabled={buyer.name === '' || buyer.email === '' || buyer.phone === ''}>Comprar</button>
+                </form>
+            </div>
+            )
+        }
+        </>
     )
 }
